Fix for CORE-2659
Minimum fix needed to solve the issue with control file generation
(CORE-2659).  Considered removing quotes from the column string
altogether, but decided against it as there could be workflows that
data coordinators have that rely on it.  As such, made this minimal
change to get us back to parity with 1.0