Modified the control file selection UI to disambiguate between choosing and editing the control file.
 * You may choose the control file by browsing for it or generating it.
 * You may then view or edit the the choosen control file.

If working from an .sij file and the control file content differs from
the control file found in the path, a warning will be issued in the UI
and the control file at the path will be used.  If running in headless
mode an error is thrown and the the .sij file will need to be cleaned
up.