Bug: Issues found during 7/22/2015 Bug Bash
Updates our control file and csv models to address issues introduced by
column names changing in the control file model.  The columns in the control
file model either are dummy placeholders, or the names of the fields in the
dataset.

Also introduces a minimum size to the advanced options panel to prevent it
from getting squished when there are a lot of columns

QA: Verified it against a 182 column dataset.  Verified that the bugs were fixed.
Added tests

Reviewed-By: TBD