Bug: Preview pipe delimited fields in the map fields dialog
If the file contained a delimiter that wasn't the default ',' but one of the fields
contained a comma (e.g. 12345|foo,bar) then we would detect a different
number of columns in each row, causing the dialog to fail to render.

This fix simply substitutes "n/a" in the preview (but not the data)
to allow the dialog to render and the customer to fix the delimiter.

Reviewed-by: TBD

QA: Verified with recent NYC upload