EN-507 - Fix networking issues with DataSync 1.6
There are a number of customers who are reporting networking
issues with DataSync 1.6.  In particular, when you click "Map
fields," the client does a bad SSL handshake and the client's firewall
shuts down the connection.  It does not repro when using
DataSync from the command line.

The root cause of this appears to be soda-java's use of the Jersey
client. Since this only repros when using Soda-java, I'm simply changing
this flow to use the HttpUtility which we know already works.  FWIW,
I did try and upgrade soda-java to Jersey 2.22 and use the same
HTTPClient.  However, it continued to hit the same renegotiation issues.

QA: Verified with Peter Moore that it fixes customer issue.  Regression tests