Update control-config.md
Updated the second reference to https://socrata.github.io/datasync/resources/faq-common-problems#how-do-i-find-the-api-field-names-for-my-columns.html.

Truth be told, I am a little lost in the relative references here so may not have gotten this quite right but you have datasync/resources/ duplicated in the currently live version, such that it returns a 404.  So, if I got this wrong, feel free to reject this pull request and just take this as a bug report.