EN-49082: rework column-matching
As a result of the NBE change we wanted to be a little fuzzy on matching
CSV/dataset columns, but this turned out to be too eager.  Now that the
NBE is everywhere, let's only do that if we're dealing with a job file
from before that change, where "before that change" is defined as "not
having a flag that says it's from after that change".